JUnit for order table impl!!!

+++ src/DAO/OrderTable.java
@@ -11,7 +11,7 @@ public interface OrderTable {
 	public int AddOrder(Order order);
 	public int GetCount();
 	public List<Order> GetAllOrders();
-	public List<Order> GetOrderForMatching(String order_category, Double quantity);
+	public List<Order> GetOrderForMatching(String order_category, int quantity);
 	public int UpdateOrderByOrderId(Order order);
 	public int GetCountUser(double user_id);
 }
